requirements.txt
# ==============================================
# ZeroDay-DRL - Required Packages
# ==============================================
#
# HOW TO INSTALL:
#   Windows:  pip install -r requirements.txt
#   Linux:    pip3 install -r requirements.txt
#
# ==============================================

# Deep Learning (PyTorch)
torch>=2.0.0
numpy>=1.24.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization (Charts and Graphs)
matplotlib>=3.7.0
seaborn>=0.12.0

# Graphical User Interface
customtkinter>=5.2.0
pillow>=10.0.0

# Progress Bars
tqdm>=4.65.0

# Configuration Files
pyyaml>=6.0
